requirements.txt
ultralytics>=8.0
torch>=2.0
torchvision>=0.15
opencv-python>=4.8
deep-sort-realtime>=1.3
numpy>=1.24
Pillow>=10.0
pygame>=2.5
fastapi>=0.104
uvicorn>=0.24
python-dotenv>=1.0
requests>=2.31
